apt: Provide markdown headers in the changelog
The `deb-changelog` structure is already very markdown-friendly, as
it uses intented bullet marks ('*', '-', '+') to list the changes.
(https://manpages.debian.org/stretch/dpkg-dev/deb-changelog.5.en.html)

On the processed changelog text, format the apt package version and
group titles as markdown headers, so that it can be directly rendered
as markdown text.